Move bit calculations to ASTContext
